requirements.txt
git+https://github.com/Rapptz/discord.py
pokedex.py
PyGithub
dateparser
gcsa
canvasapi
pyowm
Py-Dictionary
wolframalpha
requests
google-search-results
art
numpy
ollama
aiohttp